LayoutTests/imported/w3c/web-platform-tests/payment-request/delegate-request.https.sub-expected.txt

Verifies that PaymentRequest.show() call from a cross-origin subframe without user activation works if and only if the top frame has user activation and it delegates the capability to the subframe.


PASS Payment-request from a subframe fails without delegation when the top frame has no user activation
PASS Payment-request from a subframe fails without delegation when the top frame has user activation
FAIL Payment-request from a subframe succeeds with delegation when the top frame has user activation assert_equals: expected "success" but got "failure"
PASS Payment-request from a subframe fails with delegation when the top frame has no user activation
